vibrator: Use a better check for primitive effects
Some devices do have the primitive_duration node since its added
in their kernel driver [1], but do not actually support primitive
effects since they're not implemented in the devicetree (dtsi).

Check if the click primitive effect is supported; if it isn't then
its safe to assume the rest of the effects are unsupported as well.
This is necessary so that AOSP can fall back to standard effects
rather than simply failing to play the primitive effects.

[1]: https://git.codelinaro.org/clo/la/kernel/msm-5.10/-/commit/24a6bfeef28fdd6abb5161240fedd04659513b3e

Change-Id: I6f1728092074474b366afeed24a5471773067e9c